adding in a tutorial level
so i think we almost have the main level working, the last thing that i want to do is to get the 2 tutorial levels to work. they are not complicated, just showing how the players past actions effect the future actions. right now i have the past self sort of saving but it moves incorrectly. also need to add it into the overall game structure